requirements.txt
appdirs==1.4.3
asgiref==3.3.0
attrs==20.3.0
autobahn==20.7.1
Automat==20.2.0
CacheControl==0.12.6
certifi==2020.11.8
cffi==1.14.3
channels==3.0.2
chardet==3.0.4
cloudinary==1.24.0
colorama==0.4.3
constantly==15.1.0
contextlib2==0.6.0
cryptography==3.2.1
daphne==3.0.1
defusedxml==0.6.0
distlib==0.3.0
distro==1.4.0
Django==3.1.3
django-cloudinary-storage==0.3.0
django-cors-headers==3.5.0
django-oauth-toolkit==1.3.3
djangorestframework==3.12.1
drf-social-oauth2==1.0.8
gunicorn==20.0.4
html5lib==1.0.1
hyperlink==20.0.1
idna==2.10
incremental==17.5.0
ipaddr==2.2.0
lockfile==0.12.2
msgpack==0.6.2
oauthlib==3.1.0
packaging==20.3
pep517==0.8.2
Pillow==8.0.1
progress==1.5
psycopg2==2.8.6
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
PyHamcrest==2.0.2
PyJWT==1.7.1
pyOpenSSL==19.1.0
pyparsing==2.4.6
python-dotenv==0.15.0
python3-openid==3.2.0
pytoml==0.1.21
pytz==2020.4
requests==2.24.0
requests-oauthlib==1.3.0
retrying==1.3.3
service-identity==18.1.0
six==1.15.0
social-auth-app-django==4.0.0
social-auth-core==3.3.3
sqlparse==0.4.1
Twisted==20.3.0
txaio==20.4.1
urllib3==1.25.11
webencodings==0.5.1
whitenoise==5.2.0
zope.interface==5.2.0
